Move underlying source to separate file
